requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pydantic==2.5.2
crawl4ai==0.2.77
openai==1.12.0
python-dotenv==1.0.0
asyncio